--- user-manager/src/helpers/index.ts
@@ -1,26 +1,0 @@
-export { capitalizeLetter } from '@helpers/capitalizeLetter';
-export { convertToDataURL } from '@helpers/convertURL';
-export { getRandomColor } from '@helpers/getRandomColor';
-export { formatDate } from '@helpers/formatDate';
-export {
-  isEmailValid,
-  isFullNameValid
-} from '@helpers/validation';
-export { fetcher } from '@helpers/fetcher';
-export {
-  filterItem,
-  filterUsersOfRule,
-  filterRolesOfRule,
-  filterUsersOfRole,
-  filterRulesOfRole,
-  filterRulesOfUser,
-  filterRolesOfUser
-} from '@helpers/filterData';
-export { highlightKeyword } from '@helpers/highlightKeyword';
-export { generateNewUser, generateNewRole } from '@helpers/generateNewObj';
-export { isItemAssignedToUser, isItemAssignedToRole } from '@helpers/assign';
-export {
-  findUserItemId,
-  findRoleItemId,
-  findSelectedIndex
-} from '@helpers/findItem';
